slack_bot/requirements.txt
slack_bolt[async]>=1.18.0
slack_sdk>=3.27.0
supabase-py>=2.3.0
python-dotenv>=1.0
apscheduler>=3.10.4
fastapi>=0.111.0
uvicorn>=0.29.0 